# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation for LIWACH system",
        "description": "API Documentation for LIWACH system",
        "contact": {
            "email": "edenabdisa90@yahoo.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://liwachapi.herokuapp.com/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/message": {
            "get": {
                "tags": [
                    "Message"
                ],
                "summary": "Get list of message",
                "description": "Returns list of message",
                "operationId": "getMessageList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Message"
                ],
                "summary": "Store new Message",
                "description": "Returns Message data",
                "operationId": "storeMessage",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MessageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/message/{id}": {
            "put": {
                "tags": [
                    "Message"
                ],
                "summary": "Update existing Message",
                "description": "Returns updated Message data",
                "operationId": "updateMessage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MessageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Message"
                ],
                "summary": "Delete existing project",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteProject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Message": {
                "title": "Messages",
                "description": "Message model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "big integer",
                        "format": "int64",
                        "example": 1
                    },
                    "content": {
                        "title": "Text Content",
                        "description": "The text message",
                        "type": "string",
                        "example": "Hello, I want this item"
                    },
                    "type": {
                        "title": "Type",
                        "description": "It identified if the content is text or an image path",
                        "type": "string",
                        "example": "text or picture"
                    },
                    "chat_id": {
                        "title": "Chat Id",
                        "description": "The token gotten when the item was requested. Identifies the chat ID(token in company case) two user may have many chat id if they are bartering many item between them",
                        "type": "string", 
                        "example": "randomcreateduniquetoken"
                    },
                    "sender_id": {
                        "title": "Sender user id",
                        "description": "Sender user id",
                        "type": "big integer", 
                        "example": 4
                    },
                    "timestamp": {
                        "title": "Timestamp",
                        "description": "This is the timestamp that we got when this data is added.",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                }
            },
            "MessageRequest": {
                "title": "Messages",
                "description": "Message model",
                "properties": {
                    "content": {
                        "title": "Text Content",
                        "description": "The text message",
                        "type": "string",
                        "example": "Hello, I want this item"
                    },
                    "type": {
                        "title": "Type",
                        "description": "It identified if the content is text or an image path",
                        "type": "string",
                        "example": "text or picture"
                    },
                    "chat_id": {
                        "title": "Chat Id",
                        "description": "The token gotten when the item was requested. Identifies the chat ID(token in company case) two user may have many chat id if they are bartering many item between them",
                        "type": "string", 
                        "example": "randomcreateduniquetoken"
                    },
                    "sender_id": {
                        "title": "Sender user id",
                        "description": "Sender user id",
                        "type": "big integer", 
                        "example": 4
                    }
                }
            },
...
        }
    }
}